+++ app/components/add_button_component.html.erb
