--- modules/team_planner/app/components/team_planner/add_button_component.html.erb
@@ -1,9 +1,0 @@
-<li class="<%= li_css_class %>">
-  <%= link_to dynamic_path,
-              class: link_css_class,
-              title: title,
-              aria_label: aria_label do %>
-    <%= icon %>
-    <%= label %>
-  <% end %>
-</li>
